# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: College Fest Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      parameters:
        - name: college_id
          in: query
          required: false
          schema:
            type: integer
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [upcoming, completed]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"

    post:
      operationId: createEvent
      tags: [events]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/CreateEventRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /events/{eventId}:
    get:
      operationId: getEvent
      tags: [events]
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updateEvent
      tags: [events]
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/CreateEventRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteEvent
      tags: [events]
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /events/{eventId}/photos:
    get:
      operationId: getEventPhotos
      tags: [photos]
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EventPhoto"

    post:
      operationId: addEventPhoto
      tags: [photos]
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/AddPhotoRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EventPhoto"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /events/{eventId}/photos/{photoId}:
    delete:
      operationId: deleteEventPhoto
      tags: [photos]
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
        - name: photoId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /events/{eventId}/reviews:
    get:
      operationId: getEventReviews
      tags: [reviews]
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"

    post:
      operationId: addEventReview
      tags: [reviews]
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddReviewRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    Event:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        venue:
          type: string
        eventDate:
          type: string
          format: date-time
        registrationDeadline:
          type: string
          format: date-time
        maxParticipants:
          type: integer
          nullable: true
        registeredCount:
          type: integer
        thumbnailUrl:
          type: string
          nullable: true
        parentEventId:
          type: integer
          nullable: true
        previousEventId:
          type: integer
          nullable: true
        eventStatus:
          type: string
          enum: [upcoming, completed]
        collegeId:
          type: integer
        collegeName:
          type: string
        averageRating:
          type: number
          nullable: true
        reviewCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, title, description, category, venue, eventDate, registrationDeadline, collegeId, collegeName, registeredCount, eventStatus, reviewCount, createdAt]
    EventDetail:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        venue:
          type: string
        eventDate:
          type: string
          format: date-time
        registrationDeadline:
          type: string
          format: date-time
        maxParticipants:
          type: integer
          nullable: true
        registeredCount:
          type: integer
        thumbnailUrl:
          type: string
          nullable: true
        parentEventId:
          type: integer
          nullable: true
        previousEventId:
          type: integer
          nullable: true
        eventStatus:
          type: string
          enum: [upcoming, completed]
        collegeId:
          type: integer
        collegeName:
          type: string
        averageRating:
          type: number
          nullable: true
        reviewCount:
          type: integer
        photos:
          type: array
          items:
            $ref: "#/components/schemas/EventPhoto"
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
        subEvents:
          type: array
          items:
            $ref: "#/components/schemas/Event"
        createdAt:
          type: string
          format: date-time
      required: [id, title, description, category, venue, eventDate, registrationDeadline, collegeId, collegeName, registeredCount, eventStatus, reviewCount, photos, reviews, createdAt]
    CreateEventRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        venue:
          type: string
        eventDate:
          type: string
          format: date-time
        registrationDeadline:
          type: string
          format: date-time
        maxParticipants:
          type: integer
          nullable: true
        thumbnail:
          type: string
          format: binary
          nullable: true
        parentEventId:
          type: integer
          nullable: true
        previousEventId:
          type: integer
          nullable: true
      required: [title, description, category, venue, eventDate, registrationDeadline]
...
    EventPhoto:
      type: object
      properties:
        id:
          type: integer
        eventId:
          type: integer
        photoUrl:
          type: string
        caption:
          type: string
          nullable: true
        uploadedAt:
          type: string
          format: date-time
      required: [id, eventId, photoUrl, uploadedAt]
    AddPhotoRequest:
      type: object
      properties:
        photos:
          type: array
          items:
            type: string
            format: binary
        caption:
          type: string
          nullable: true
      required: [photos]
    Review:
      type: object
      properties:
        id:
          type: integer
        eventId:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
        rating:
          type: integer
        review:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, eventId, studentId, studentName, rating, createdAt]
    AddReviewRequest:
      type: object
      properties:
        rating:
          type: integer
        review:
          type: string
          nullable: true
      required: [rating]