# lib/api-zod/src/generated/types/addPhotoRequest.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Api
 * College Fest Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface AddPhotoRequest {
  photos: Blob[];
  caption?: string | null;
}