# lib/api-zod/src/generated/types/addReviewRequest.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Api
 * College Fest Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface AddReviewRequest {
  rating: number;
  review?: string | null;
}