# lib/api-zod/src/generated/types/review.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Api
 * College Fest Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface Review {
  id: number;
  eventId: number;
  studentId: number;
  studentName: string;
  rating: number;
  review?: string | null;
  createdAt: Date;
}